refactor to make addEvent a larger utility function that contains other functions
