+++ server/routes.js
@@ -1,7 +1,7 @@
 const path = require('path');
 const router = require('express').Router();
 const User = require('./db/controllers/userController.js');
-const InsertEvent = require('./db/controllers/eventApi.js');
+const addEvent = require('./utility/addEvent');
 const CalendarEvents = require('./db/controllers/eventController.js');
 const GoogleAuthUrl = require('./setup/googleOAuth').url;
 const GoogleMaps = require('./utility/googleMaps');
@@ -14,8 +14,9 @@ router.get('/auth', function(req, res) {
 });
 
 router.get('/authCallback', User.createUser);
+// this is tricky to fix, probably want to serve up a static page
 
-router.post('/calendar/addEvent', InsertEvent.postEventToApi);
+router.post('/calendar/addEvent', addEvent);
 router.get('/calendar/getEvent', CalendarEvents.retrieveDayEvent);
 router.get('/calendar/getAllEvents', getAllEventsFromCalendar);
